Add React router navigation and update various components
Enhanced the City, CountryList Form, and Map components to make use of React Router's navigation features. This allows us to direct the user to different views based on the application's URL. Also, updated the BrowserRouter import in App.jsx to include Navigate, and included routing for the City and Form components. As a result, we removed the use of hardcoded values in the City component and replaced it with dynamic routing parameters and search parameters for better flexibility. Also added a new Button component, and utilized it in the Form's back button, along with useNavigate hook.