Replace useState with useReducer in CitiesContext
The useState function in the CitiesContext component has been replaced with useReducer for better state organization and management. This change includes the introduction of a reducer function and initial state. It also refactors the component methods (fetch, create and delete) to use the dispatch function from useReducer. This provides a more predictable state transition and makes the application easier to test and debug.